Refactor autoblocker handling by centralizing logic in processAutoblockerResults function
